// src/types/stamps.ts
export type LocationData = {
    name: string;
    stampImage: string;
    image: string;
    intro: string;
    collected: boolean;
  };

  export type SocialData = {
    name: string;
    stampImage: string;
    image: string;
    description: string;
    timestamp: string;
    collected: boolean;
    count: number;
  };